+++ frontend/src/components/messstelle/optionsmenue/panels/content/FahrzeugPanelVerkehrsartenContent.vue
@@ -188,7 +188,7 @@
 
 <script setup lang="ts">
 import Fahrzeug from "@/types/enum/Fahrzeug";
-import MessstelleOptionsDTO from "@/types/messung/MessstelleOptionsDTO";
+import MessstelleOptionsDTO from "@/types/messstelle/MessstelleOptionsDTO";
 import { computed, ref, watch } from "vue";
 import PanelHeader from "@/components/common/PanelHeader.vue";
 import { useFahrzeugPanelTools } from "@/components/messstelle/optionsmenue/composable/fahrzeugPanelTools";
