+++ frontend/src/types/messstelle/FahrzeugOptions.ts
