+++ frontend/src/types/messstelle/MessquerschnittInfoDTO.ts
